feat(cli): list filtered Core events
