+++ crates/ctx-cli/src/commands/mod.rs
@@ -1,9 +1,9 @@
 pub(crate) mod blame;
 pub(crate) mod doctor;
-pub(crate) mod export;
 pub(crate) mod import;
 pub(crate) mod index;
 mod index_dashboard;
+pub(crate) mod list;
 pub(crate) mod locate;
 pub(crate) mod search;
 pub(crate) mod setup;
