+++ crates/ctx-cli/src/local_usage/mod.rs
@@ -326,17 +326,17 @@ impl CliUsage {
                 match &args.target {
                     ShowTarget::Session(_) => Some("show_session"),
                     ShowTarget::Event(_) => Some("show_event"),
-                    ShowTarget::Events(_) => Some("show_event"),
                 },
                 TargetType::NotApplicable,
             ),
+            CommandRoot::List(_) => (Some("show_event"), TargetType::NotApplicable),
             CommandRoot::Locate(_) => (Some("locate"), TargetType::NotApplicable),
             CommandRoot::Search(_) => (Some("search"), TargetType::NotApplicable),
             CommandRoot::Pro(args) => (
                 Some(args.local_usage_operation()),
                 TargetType::NotApplicable,
             ),
-            CommandRoot::Referral(_) | CommandRoot::Export(_) => (None, TargetType::NotApplicable),
+            CommandRoot::Referral(_) => (None, TargetType::NotApplicable),
             CommandRoot::Blame(_) => (Some("blame"), TargetType::NotApplicable),
             CommandRoot::Docs(_) => (Some("docs"), TargetType::NotApplicable),
             CommandRoot::Integrations(_) => (Some("integrations"), TargetType::NotApplicable),
